ui(operations): extract DatabaseTab into its own module
Lifts the database tab body (stat row, vault file path, tables list,
schedule panel, optimize/vacuum/reindex/integrity-check action
panels) plus its supporting helpers (DatabaseActions,
ScheduledMaintenanceCard, TablesTable) out of GroveMaintenance.tsx.
Behavior is unchanged.
